+++ @remirror/core/src/portal-container.ts
@@ -20,6 +20,10 @@ interface Events {
 export type PortalList = ReadonlyArray<[HTMLElement, MountedPortal]>;
 export type PortalMap = Map<HTMLElement, MountedPortal>;
 
+/**
+ * The node view portal container keeps track of all the portals which have been added by react to render
+ * the node views in the editor.
+ */
 export class NodeViewPortalContainer {
   public portals: Map<HTMLElement, MountedPortal> = new Map();
   public context!: Component<NodeViewPortalComponentProps>;